+++ application/views/pages/changelog.php
@@ -4,6 +4,11 @@
             <a href="<?php echo base_url(); ?>">Retour à l'accueil</a>
         </div>
         <div class="version mb-6">
+            <h2 class="mb-4">Version 0.2.0 - 18/04/2020</h2>
+            <p>Nouvelle section « prix indicatif » sur la page des consommables de l'outil</p>
+            <p>Nouvelle sections « ressources » sur la page info de l'outil</p>
+        </div>
+        <div class="version mb-6">
             <h2 class="mb-4">Version 0.1.0 - 17/04/2020</h2>
             <p>Nouveau header</p>
             <p>Nouvelle mise en page pour les pages outils</p>