+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## Version 0.2.0 - 18/04/2020
+
+New price input on tool's consumables page
+New ressources section on tool's info page
+
 ## Version 0.1.0 - 17/04/2020
 
 New header layout